Added tests for OpenCL expression writer handling of conditions. See #602.
